remove cyclic dependency between chatSetup and chatViewPane
fyi @roblourens
